[#45] fix: RankTitleList 초기화를 StatDrop컴포넌트에서 하게 변경
